+++ vars/helm.groovy
@@ -437,8 +437,7 @@ void show(Map config) {
   // input checking
   config.bin = config.bin ?: 'helm'
   assert config.chart instanceof String : 'The required parameter "chart" was not set.'
-  assert config.info instanceof String : 'The required parameter "info" was not set.'
-  assert (['all', 'chart', 'readme', 'values']).contains(config.info) : "The info parameter must be one of all, chart, readme, or values."
+  assert (['all', 'chart', 'crds', 'readme', 'values']).contains(config.info) : "The info parameter must be one of all, chart, crds, readme, or values."
 
   // show chart info
   try {