+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `force` parameter to `install`, `repo`, `rollback`, and `upgrade` methods.
 - Add `withSubcharts` parameter to `lint` method.
 - Add `hooks` and `recreatePods` parameters to `rollback` method.
+- Support `crds` for `show` method.
 
 **Packer**
 - Add `pluginsInstall` and `pluginsRemove` methods.