+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-### 2.2.0 (Next)
+### 2.1.2 (Next)
 **Helm**
 - Add `force` parameter to `install`, `repo`, `rollback`, and `upgrade` methods.
 - Add `withSubcharts` parameter to `lint` method.
 - Add `hooks` and `recreatePods` parameters to `rollback` method.
+- Add `outputFormat` parameter to `status` method.
 - Support `crds` for `show` method.
 
 **Packer**